+++ scanpipe/views.py
@@ -25,7 +25,6 @@ import io
 import json
 import operator
 from collections import Counter
-from collections import namedtuple
 from contextlib import suppress
 from pathlib import Path
 
@@ -64,12 +63,14 @@ from scanpipe.filters import DependencyFilterSet
 from scanpipe.filters import ErrorFilterSet
 from scanpipe.filters import PackageFilterSet
 from scanpipe.filters import ProjectFilterSet
+from scanpipe.filters import RelationFilterSet
 from scanpipe.filters import ResourceFilterSet
 from scanpipe.forms import AddInputsForm
 from scanpipe.forms import AddPipelineForm
 from scanpipe.forms import ArchiveProjectForm
 from scanpipe.forms import ProjectForm
 from scanpipe.forms import ProjectSettingsForm
+from scanpipe.models import CodebaseRelation
 from scanpipe.models import CodebaseResource
 from scanpipe.models import DiscoveredDependency
 from scanpipe.models import DiscoveredPackage
@@ -81,6 +82,10 @@ from scanpipe.pipes import count_group_by
 from scanpipe.pipes import output
 
 scanpipe_app = apps.get_app_config("scanpipe")
+
+
+# Cancel the default ordering for better performances
+unordered_resources = CodebaseResource.objects.order_by()
 
 
 LICENSE_CLARITY_FIELDS = [
@@ -1119,8 +1124,7 @@ class DiscoveredPackageListView(
     prefetch_related = [
         Prefetch(
             "codebase_resources",
-            # Cancel the default ordering for better performances
-            queryset=CodebaseResource.objects.order_by().only("path", "name"),
+            queryset=unordered_resources.only("path", "name"),
         ),
     ]
     table_columns = [
@@ -1200,6 +1204,9 @@ class DiscoveredDependencyListView(
         },
     ]
 
+    def get_queryset(self):
+        return super().get_queryset().order_by("dependency_uid")
+
 
 class ProjectErrorListView(
     ConditionalLoginRequired,
@@ -1220,91 +1227,46 @@ class ProjectErrorListView(
     ]
 
 
-RelationRow = namedtuple(
-    "RelationRow",
-    field_names=["to_resource", "status", "map_type", "score", "from_resource"],
-)
-
-
 class CodebaseRelationListView(
     ConditionalLoginRequired,
     ProjectRelatedViewMixin,
+    PrefetchRelatedViewMixin,
     TableColumnsMixin,
     ExportXLSXMixin,
     PaginatedFilterView,
 ):
-    model = CodebaseResource
-    filterset_class = ResourceFilterSet
+    model = CodebaseRelation
+    filterset_class = RelationFilterSet
     template_name = "scanpipe/relation_list.html"
+    prefetch_related = [
+        Prefetch(
+            "to_resource",
+            queryset=unordered_resources.only("path", "is_text", "status"),
+        ),
+        Prefetch(
+            "from_resource",
+            queryset=unordered_resources.only("path", "is_text", "status"),
+        ),
+    ]
     paginate_by = settings.SCANCODEIO_PAGINATE_BY.get("relation", 100)
     table_columns = [
-        {
-            "field_name": "path",
-            "label": "To resource",
-        },
+        "to_resource",
         {
             "field_name": "status",
             "filter_fieldname": "status",
         },
         {
-            "field_name": "related_from__map_type",
-            "label": "Map type",
-            "filter_fieldname": "relation_map_type",
+            "field_name": "map_type",
+            "filter_fieldname": "map_type",
         },
-        {
-            "field_name": "related_from__from_resource__path",
-            "label": "From resource",
-        },
+        "from_resource",
     ]
 
-    def get_queryset(self):
-        return (
-            super()
-            .get_queryset()
-            .files()
-            .to_codebase()
-            .prefetch_related("related_from__from_resource__project")
-        )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["relation_count"] = context["filter"].qs.has_relation().count()
-        return context
-
-    @staticmethod
-    def get_rows(qs):
-        for resource in qs:
-            relations = resource.related_from.all()
-            if not relations:
-                yield RelationRow(resource.path, resource.status, "", "", "")
-            else:
-                for relation in resource.related_from.all():
-                    score = relation.extra_data.get("path_score", "")
-                    if diff_ratio := relation.extra_data.get("diff_ratio", ""):
-                        score += f" diff_ratio: {diff_ratio}"
-                    yield RelationRow(
-                        resource.path,
-                        resource.status,
-                        relation.map_type,
-                        score,
-                        relation.from_resource.path,
-                    )
-
-    def export_xlsx_file_response(self):
-        filtered_qs = self.filterset.qs
-        output_file = io.BytesIO()
-
-        with xlsxwriter.Workbook(output_file) as workbook:
-            output._add_xlsx_worksheet(
-                workbook=workbook,
-                worksheet_name="RELATIONS",
-                rows=self.get_rows(qs=filtered_qs),
-                fields=RelationRow._fields,
-            )
-
-        filename = f"{self.project.name}_{self.model._meta.model_name}.xlsx"
-        output_file.seek(0)
-        return FileResponse(output_file, as_attachment=True, filename=filename)
+    def get_filterset_kwargs(self, filterset_class):
+        """Add the project in the filterset kwargs for computing status choices."""
+        kwargs = super().get_filterset_kwargs(filterset_class)
+        kwargs.update({"project": self.project})
+        return kwargs
 
 
 class CodebaseResourceDetailsView(